agents/requirements.txt
uvicorn
fastapi
python-dotenv
requests
ic-py
uagents